MCP_weather/mcp_client/requirements.txt
mcp
openai
python-dotenv
fastapi
uvicorn[standard]
jinja2
requests 